# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "wavynote",
            "email": "chelseafandev@gmail.com"
        }
    },
    "paths": {
...
        "/search/folder": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "전체 폴더를 대상으로 노트 내용 검색",
                "tags": [
                    "나의노트 페이지"
                ],
                "summary": "전체 폴더를 대상으로 노트 내용 검색",
                "parameters": [
                    {
                        "type": "string",
                        "description": "target folder id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query for search",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.NoteListResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/search/opennote": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "전체 폴더를 대상으로 노트 내용 검색",
                "tags": [
                    "오픈노트 페이지"
                ],
                "summary": "전체 폴더를 대상으로 노트 내용 검색",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query for search",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.OpenNoteListResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/search/top": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "전체 폴더를 대상으로 노트 내용 검색",
                "tags": [
                    "나의노트 페이지"
                ],
                "summary": "전체 폴더를 대상으로 노트 내용 검색",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query for search",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.NoteListResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "restapi.NoteListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restapi.NoteSimpleInfo"
                    }
                }
            }
        },
        "restapi.NoteSimpleInfo": {
            "type": "object",
            "properties": {
                "note_id": {
                    "description": "해당 폴더에 존재하는 노트의 고유 id 값",
                    "type": "string",
                    "example": "09d05df1-2958-4a3d-b910-3b4fb079327b"
                },
                "preview": {
                    "description": "해당 폴더에 존재하는 노트의 본문 미리보기(글자수 제한)",
                    "type": "string",
                    "example": "This is the main text of my first wavy note."
                },
                "title": {
                    "description": "해당 폴더에 존재하는 노트의 제목",
                    "type": "string",
                    "example": "my first note"
                }
            }
        },
        "restapi.OpenNoteListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restapi.OpenNoteSimpleInfo"
                    }
                }
            }
        },
        "restapi.OpenNoteSimpleInfo": {
            "type": "object",
            "properties": {
                "conversation_id": {
                    "type": "string",
                    "example": "1afc571d-61bf-4cef-95ce-ab791f999297"
                },
                "note_id": {
                    "type": "string",
                    "example": "09d05df1-2958-4a3d-b910-3b4fb079327b"
                },
                "post_at": {
                    "type": "string",
                    "example": "2023-12-19 15:55:48"
                },
                "preview": {
                    "type": "string",
                    "example": "This is the main text of my first wavy note."
                },
                "title": {
                    "type": "string",
                    "example": "my first note"
                }
            }
        },
...
        "restapi.Response400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "Bad request"
                }
            }
        },
        "restapi.Response401": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "msg": {
                    "type": "string",
                    "example": "Fail to parse Authorization header"
                }
            }
        },
        "restapi.Response404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "msg": {
                    "type": "string",
                    "example": "Not exist the target resource in our server"
                }
            }
        },
        "restapi.Response500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}